+++ lib/flutter_redux.dart
@@ -3,7 +3,6 @@ library flutter_redux;
 import 'dart:async';
 
 import 'package:flutter/widgets.dart';
-import 'package:meta/meta.dart';
 import 'package:redux/redux.dart';
 
 /// Provides a Redux [Store] to all descendants of this Widget. This should
@@ -15,12 +14,10 @@ class StoreProvider<S> extends InheritedWidget {
   /// Create a [StoreProvider] by passing in the required [store] and [child]
   /// parameters.
   const StoreProvider({
-    Key key,
-    @required Store<S> store,
-    @required Widget child,
-  })  : assert(store != null),
-        assert(child != null),
-        _store = store,
+    Key? key,
+    required Store<S> store,
+    required Widget child,
+  })   : _store = store,
         super(key: key, child: child);
 
   /// A method that can be called by descendant Widgets to retrieve the Store
@@ -76,7 +73,7 @@ class StoreProvider<S> extends InheritedWidget {
         ? context.dependOnInheritedWidgetOfExactType<StoreProvider<S>>()
         : context
             .getElementForInheritedWidgetOfExactType<StoreProvider<S>>()
-            ?.widget) as StoreProvider<S>;
+            ?.widget) as StoreProvider<S>?;
 
     if (provider == null) throw StoreProviderError<StoreProvider<S>>();
 
@@ -137,22 +134,38 @@ typedef IgnoreChangeTest<S> = bool Function(S state);
 
 /// A function that will be run on State change, before the build method.
 ///
-/// This function is passed the `ViewModel`, and if `distinct` is `true`,
-/// it will only be called if the `ViewModel` changes.
+/// This function is passed the previous and current `ViewModel`, and if
+/// `distinct` is `true`, it will only be called when the `ViewModel` changes.
 ///
 /// This is useful for making calls to other classes, such as a
 /// `Navigator` or `TabController`, in response to state changes.
 /// It can also be used to trigger an action based on the previous
 /// state.
+///
+/// ```dart
+/// StoreConnector<String, String>(
+///   converter: (store) => store.state,
+///   onWillChange: (prev, vm) {
+///     if (prev != vm) {
+///       ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+///         content: Text(vm),
+///       ));
+///     }
+///   },
+///   builder: (context, vm) {
+///     return Text(vm);
+///   },
+/// );
+/// ```
 typedef OnWillChangeCallback<ViewModel> = void Function(
-  ViewModel previousViewModel,
+  ViewModel? previousViewModel,
   ViewModel newViewModel,
 );
 
 /// A function that will be run on State change, after the build method.
 ///
-/// This function is passed the `ViewModel`, and if `distinct` is `true`,
-/// it will only be called if the `ViewModel` changes.
+/// This function is passed the previous and current `ViewModel`, and if
+/// `distinct` is `true`, it will only be called when the `ViewModel` changes.
 ///
 /// This can be useful for running certain animations after the build is
 /// complete.
@@ -160,7 +173,28 @@ typedef OnWillChangeCallback<ViewModel> = void Function(
 /// Note: Using a [BuildContext] inside this callback can cause problems if
 /// the callback performs navigation. For navigation purposes, please use
 /// an [OnWillChangeCallback].
-typedef OnDidChangeCallback<ViewModel> = void Function(ViewModel viewModel);
+///
+/// ```dart
+/// StoreConnector<int, int>(
+///   converter: (store) => store.state,
+///   onDidChange: (prev, vm) {
+///     if (prev != vm) {
+///       myScrollController.animateTo(200);
+///     }
+///   },
+///   builder: (context, vm) {
+///     return ListView.builder(
+///       controller: myScrollController,
+///       itemCount: vm,
+///       builder: (context, index) => Text('$index'),
+///     );
+///   },
+/// );
+/// ```
+typedef OnDidChangeCallback<ViewModel> = void Function(
+  ViewModel? previousViewModel,
+  ViewModel viewModel,
+);
 
 /// A function that will be run after the Widget is built the first time.
 ///
@@ -201,7 +235,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
   ///
   /// This can be useful for dispatching actions that fetch data for your Widget
   /// when it is first displayed.
-  final OnInitCallback<S> onInit;
+  final OnInitCallback<S>? onInit;
 
   /// A function that will be run when the StoreConnector is removed from the
   /// Widget Tree.
@@ -210,7 +244,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
   ///
   /// This can be useful for dispatching actions that remove stale data from
   /// your State tree.
-  final OnDisposeCallback<S> onDispose;
+  final OnDisposeCallback<S>? onDispose;
 
   /// Determines whether the Widget should be rebuilt when the Store emits an
   /// onChange event.
@@ -230,7 +264,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
   /// If you ignore a change, and the framework needs to rebuild the Widget, the
   /// [builder] function will be called with the latest [ViewModel] produced by
   /// your [converter] function.
-  final IgnoreChangeTest<S> ignoreChange;
+  final IgnoreChangeTest<S>? ignoreChange;
 
   /// A function that will be run on State change, before the Widget is built.
   ///
@@ -240,7 +274,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
   /// This can be useful for imperative calls to things like Navigator,
   /// TabController, etc. This can also be useful for triggering actions
   /// based on the previous state.
-  final OnWillChangeCallback<ViewModel> onWillChange;
+  final OnWillChangeCallback<ViewModel>? onWillChange;
 
   /// A function that will be run on State change, after the Widget is built.
   ///
@@ -253,7 +287,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
   /// Note: Using a [BuildContext] inside this callback can cause problems if
   /// the callback performs navigation. For navigation purposes, please use
   /// [onWillChange].
-  final OnDidChangeCallback<ViewModel> onDidChange;
+  final OnDidChangeCallback<ViewModel>? onDidChange;
 
   /// A function that will be run after the Widget is built the first time.
   ///
@@ -262,7 +296,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
   ///
   /// This can be useful for starting certain animations, such as showing
   /// Snackbars, after the Widget is built the first time.
-  final OnInitialBuildCallback<ViewModel> onInitialBuild;
+  final OnInitialBuildCallback<ViewModel>? onInitialBuild;
 
   /// Create a [StoreConnector] by passing in the required [converter] and
   /// [builder] functions.
@@ -271,9 +305,9 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
   /// modify the behavior of the StoreConnector. Please see the documentation
   /// for each option for more info.
   const StoreConnector({
-    Key key,
-    @required this.builder,
-    @required this.converter,
+    Key? key,
+    required this.builder,
+    required this.converter,
     this.distinct = false,
     this.onInit,
     this.onDispose,
@@ -282,9 +316,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
     this.onWillChange,
     this.onDidChange,
     this.onInitialBuild,
-  })  : assert(builder != null),
-        assert(converter != null),
-        super(key: key);
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -325,7 +357,7 @@ class StoreBuilder<S> extends StatelessWidget {
   ///
   /// This can be useful for dispatching actions that fetch data for your Widget
   /// when it is first displayed.
-  final OnInitCallback<S> onInit;
+  final OnInitCallback<S>? onInit;
 
   /// A function that will be run when the StoreBuilder is removed from the
   /// Widget Tree.
@@ -334,14 +366,14 @@ class StoreBuilder<S> extends StatelessWidget {
   ///
   /// This can be useful for dispatching actions that remove stale data from
   /// your State tree.
-  final OnDisposeCallback<S> onDispose;
+  final OnDisposeCallback<S>? onDispose;
 
   /// A function that will be run on State change, before the Widget is built.
   ///
   /// This can be useful for imperative calls to things like Navigator,
   /// TabController, etc. This can also be useful for triggering actions
   /// based on the previous state.
-  final OnWillChangeCallback<Store<S>> onWillChange;
+  final OnWillChangeCallback<Store<S>>? onWillChange;
 
   /// A function that will be run on State change, after the Widget is built.
   ///
@@ -351,26 +383,25 @@ class StoreBuilder<S> extends StatelessWidget {
   /// Note: Using a [BuildContext] inside this callback can cause problems if
   /// the callback performs navigation. For navigation purposes, please use
   /// [onWillChange].
-  final OnDidChangeCallback<Store<S>> onDidChange;
+  final OnDidChangeCallback<Store<S>>? onDidChange;
 
   /// A function that will be run after the Widget is built the first time.
   ///
   /// This can be useful for starting certain animations, such as showing
   /// Snackbars, after the Widget is built the first time.
-  final OnInitialBuildCallback<Store<S>> onInitialBuild;
+  final OnInitialBuildCallback<Store<S>>? onInitialBuild;
 
   /// Create's a Widget based on the Store.
   const StoreBuilder({
-    Key key,
-    @required this.builder,
+    Key? key,
+    required this.builder,
     this.onInit,
     this.onDispose,
     this.rebuildOnChange = true,
     this.onWillChange,
     this.onDidChange,
     this.onInitialBuild,
-  })  : assert(builder != null),
-        super(key: key);
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -394,18 +425,18 @@ class _StoreStreamListener<S, ViewModel> extends StatefulWidget {
   final Store<S> store;
   final bool rebuildOnChange;
   final bool distinct;
-  final OnInitCallback<S> onInit;
-  final OnDisposeCallback<S> onDispose;
-  final IgnoreChangeTest<S> ignoreChange;
-  final OnWillChangeCallback<ViewModel> onWillChange;
-  final OnDidChangeCallback<ViewModel> onDidChange;
-  final OnInitialBuildCallback<ViewModel> onInitialBuild;
+  final OnInitCallback<S>? onInit;
+  final OnDisposeCallback<S>? onDispose;
+  final IgnoreChangeTest<S>? ignoreChange;
+  final OnWillChangeCallback<ViewModel>? onWillChange;
+  final OnDidChangeCallback<ViewModel>? onDidChange;
+  final OnInitialBuildCallback<ViewModel>? onInitialBuild;
 
   const _StoreStreamListener({
-    Key key,
-    @required this.builder,
-    @required this.store,
-    @required this.converter,
+    Key? key,
+    required this.builder,
+    required this.store,
+    required this.converter,
     this.distinct = false,
     this.onInit,
     this.onDispose,
@@ -424,22 +455,24 @@ class _StoreStreamListener<S, ViewModel> extends StatefulWidget {
 
 class _StoreStreamListenerState<S, ViewModel>
     extends State<_StoreStreamListener<S, ViewModel>> {
-  Stream<ViewModel> _stream;
-  ViewModel _latestValue;
-  ConverterError _latestError;
-  S _lastConvertedState;
+  late Stream<ViewModel> _stream;
+  ViewModel? _latestValue;
+  ConverterError? _latestError;
+
+  // `_latestValue!` would throw _CastError if `ViewModel` is nullable,
+  // therefore `_latestValue as ViewModel` is used.
+  // https://dart.dev/null-safety/understanding-null-safety#nullability-and-generics
+  ViewModel get _requireLatestValue => _latestValue as ViewModel;
 
   @override
   void initState() {
-    if (widget.onInit != null) {
-      widget.onInit(widget.store);
-    }
+    widget.onInit?.call(widget.store);
 
     _computeLatestValue();
 
     if (widget.onInitialBuild != null) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        widget.onInitialBuild(_latestValue);
+      WidgetsBinding.instance?.addPostFrameCallback((_) {
+        widget.onInitialBuild!(_requireLatestValue);
       });
     }
 
@@ -450,9 +483,7 @@ class _StoreStreamListenerState<S, ViewModel>
 
   @override
   void dispose() {
-    if (widget.onDispose != null) {
-      widget.onDispose(widget.store);
-    }
+    widget.onDispose?.call(widget.store);
 
     super.dispose();
   }
@@ -484,23 +515,17 @@ class _StoreStreamListenerState<S, ViewModel>
         ? StreamBuilder<ViewModel>(
             stream: _stream,
             builder: (context, snapshot) {
-              if (_latestError != null) throw _latestError;
+              if (_latestError != null) throw _latestError!;
 
               return widget.builder(
                 context,
-                _latestValue,
+                _requireLatestValue,
               );
             },
           )
         : _latestError != null
-            ? throw _latestError
-            : widget.builder(context, _latestValue);
-  }
-
-  bool _stateChanged(S state) {
-    var ifStateChanged = !identical(_lastConvertedState, widget.store.state);
-    _lastConvertedState = widget.store.state;
-    return ifStateChanged;
+            ? throw _latestError!
+            : widget.builder(context, _requireLatestValue);
   }
 
   ViewModel _mapConverter(S state) {
@@ -517,7 +542,7 @@ class _StoreStreamListenerState<S, ViewModel>
 
   bool _ignoreChange(S state) {
     if (widget.ignoreChange != null) {
-      return !widget.ignoreChange(widget.store.state);
+      return !widget.ignoreChange!(widget.store.state);
     }
 
     return true;
@@ -525,7 +550,6 @@ class _StoreStreamListenerState<S, ViewModel>
 
   void _createStream() {
     _stream = widget.store.onChange
-        .where(_stateChanged)
         .where(_ignoreChange)
         .map(_mapConverter)
         // Don't use `Stream.distinct` because it cannot capture the initial
@@ -535,21 +559,22 @@ class _StoreStreamListenerState<S, ViewModel>
         // latestValue. Important: This must be done after all other optional
         // transformations, such as ignoreChange.
         .transform(StreamTransformer.fromHandlers(
-            handleData: _handleChange, handleError: _handleError));
+          handleData: _handleChange,
+          handleError: _handleError,
+        ));
   }
 
   void _handleChange(ViewModel vm, EventSink<ViewModel> sink) {
     _latestError = null;
-
-    if (widget.onWillChange != null) {
-      widget.onWillChange(_latestValue, vm);
-    }
-
+    widget.onWillChange?.call(_latestValue, vm);
+    final previousValue = vm;
     _latestValue = vm;
 
     if (widget.onDidChange != null) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        widget.onDidChange(_latestValue);
+      WidgetsBinding.instance?.addPostFrameCallback((_) {
+        if (mounted) {
+          widget.onDidChange!(previousValue, _requireLatestValue);
+        }
       });
     }
 
@@ -599,6 +624,7 @@ class ConverterError extends Error {
   final Object error;
 
   /// The stacktrace that accompanies the [error]
+  @override
   final StackTrace stackTrace;
 
   /// Creates a ConverterError with the relevant error and stacktrace